% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRP.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read.vectors}
\alias{reexports}
\alias{write.binary.word2vec}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{wordVectors}{\code{\link[wordVectors]{read.vectors}}, \code{\link[wordVectors]{write.binary.word2vec}}}
}}

