% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRP.R
\name{stable_transform}
\alias{stable_transform}
\title{Generate the SRP encoding of a string}
\usage{
stable_transform(words, counts = NULL, standardize = T, log = T,
  length_out = 160)
}
\arguments{
\item{words}{A string or vector of strings. If standardize is TRUE, this will be tokenized for you.}

\item{counts}{The number of times each string occurs. If NULL (the default), each string is assumed to occur once.}

\item{standardize}{Whether to perform standard tokenization.}

\item{log}{Whether to use a log transform.}

\item{length_out}{The desired dimensionality}
}
\value{
A vector giving the
}
\description{
Generate the SRP encoding of a string
}
\examples{

#The two functions below return the same amount. The first takes longer to run but less time to write.
stable_transform("oh happy happy day", length_out = 10)
stable_transform(c("oh", "happy", "day"), counts = c(1, 2, 1), length_out = 10)

}
